Fixed two bugs associated with importing xml recipe files.
When recipes are exported as Gourmet XML and reimported it was
observed that the value of the yield was incorrect if it wasn't
an integer.  This was fixed by modifying the regex pattern used
to match the subparts of the yield line.

Additionally, it was found that the unit was correctly parsed in
one part of the import and then ignored and a generic value of
'serving' was instead substituted.  The logic only adds serving
if no other unit was found.  For example, if the yield is '4.75
pints' then it is recorded as '4.75 pints' and if the yield is
simply '4.75' then it is recorded as '4.75 servings'.  Previously,
both examples would have been recorded as '4.75 servings'.

Added in module tests at the bottom of importer.py.