Fixed failing mcb import test caused by previous commit.
One of the tests failed due to a key error during the import
process.  The changed conditional first fails if the key is
not present in the dictionary and then fails if it is present
but the value evaluates to False (empty string).

This is necessary because this part of the parsing logic is
used with multiple file types and sometimes the yield unit has
already been parsed and assigned and sometimes it still needs
to be assigned.