Fix yield when importing exported XML file (#210)
Fixed two bugs associated with importing XML recipe files.

When recipes are exported as Gourmet XML and re-imported it was
observed that the value of the yield was incorrect if it wasn't
an integer.  This was fixed by modifying the regex pattern used
to match the subparts of the yield line.

Additionally, it was found that the unit was correctly parsed in
one part of the import and then ignored and a generic value of
'serving' was instead substituted.  The logic only adds serving
if no other unit was found.  For example, if the yield is '4.75
pints' then it is recorded as '4.75 pints' and if the yield is
simply '4.75' then it is recorded as '4.75 servings'.  Previously,
both examples would have been recorded as '4.75 servings'.

One of the tests failed due to a key error during the import
process.  The changed conditional first fails if the key is
not present in the dictionary and then fails if it is present
but the value evaluates to False (empty string).

This is necessary because this part of the parsing logic is
used with multiple file types and sometimes the yield unit has
already been parsed and assigned and sometimes it still needs
to be assigned.

Closes #165.