Do not fail for images without a src attribute (#211)
Closes #147.

The web import in an unsupported website leads to a KeyError if
there are no images on the recipe page.  This commit adds a
precondition to the if statement in a list comprehension that checks
to make sure the dictionary key is present.  This prevents trying
to access a key which is not present in the dictionary.