Modernize commands
This ensures that editable installs keep working after the explicit `setup.py install`
call is removed. Additionally, declare the build backend in the `pyproject.toml` file.

As custom `setup.py *` calls are discouraged, move `update_i18n` to a dedicated script
and update the corresponding files as well. This ensures that the additional migration
from the (possibly unsafe) `os.system` to `subprocess.run` did not break anything as
well.

Closes #203.