Fix xml around instructions and notes (modifications).
Related to issue #77. Gourmand double escapes xml special characters
&, < and > when exporting.  This causes an ampersand (&) in the text
to become &amp;amp;.  A similar issue happens with > and <.  To combat
this an extra call to xml.sax.saxutils.unescape() was used in the
exporter class private method _write_text_.

The opposite issue is present when importing.  The special characters
needed to be escaped.  To achieve this the importer.Importer.__init__()
method is given a default do_markup value of True.  This seems to cause
the special symbols to be escaped properly when importing and saving.

These escaping issues seemed to only be present in the instructions and
modification sections of the recipe, which indicates the solution could
be found by changing code in those code paths.