Fixes import and export escaping in XML for formatting tags (e.g. <b>).
The previous commit worked to allow <, > and & to be imported and
exported since they commonly interfere with XML parsing.  Unfortunately,
it was found the fix for that broke saved formatting information
which is communicated through <b>, <u> and <i> tags, and possibly
others.

The errors with handling formatting tags and special characters at the
same time stem from improper escaping during both the export and import
process.  To fix with with exporting a single extra escape line was added
to the exporter class.  This extra escape is only applied to the text
portion of the XML document and not the tags themselves and it has the
effect of preserving the relative escape level (for lack of a better term).

During the import process both an unescape and an escape portion of the
code applied to 'instructions' and 'modifications' had to be removed.
While it might seem like these cancel out, they were applied in such a
way to abolish the relative escape level, and removing these again helps
preserve the relative escape level during the import process.

See the thread in issue #77 for more context.