Added automated test for .grmt xml file with special and formatting characters.
An extra file was added at tests/recipe_files/special_chars.grmt.
This contains the gourmetDoc xml file for a simple recipe with only
a title and instructions with complicated characters and format.

For some reason re.match did not work, probably due to the special
characters or newlines.  For this reason I added a couple functions
modelled off of the other ones, to import the file and make sure
the string strictly matches what is expected.