Refactored import_urls() and added test for rating.
Added capture_image_and_thumbnail() function to create the image
and thumbnail from the scraped recipe.

Added initialize_recipe() function to create the recipe.  Since it
is a named tuple we need to pass the image and thumbnail information
in to include it at the time of initialization.

Import_urls is still the top function to call, same arguments
accepted and return values.

To test_web_importer.py added a function that confirms when a recipe
is created with no rating value given it is returned with a rating
value of 0 as the default.  The import and object initialization
succeeds.