Fixed time parsing for ranges.
The regex used to match time ranges was slightly modified to allow capture
of the range and second number values, when present.  These changes were
in convert.py.

The timeScanner.py make_time_links() function was modified to handle both
simple and complex time ranges.  This required modifiying the approach
from a single global substitution to an iterative match and single
substitution method.  Each match is determined to be a single time or
a time range and the substitution used is different.  Overall the output
format of make_time_links() remains unchanged.

Automated tests were added to test_time_scanner.py.

See issue #231 for more context.