Adds a cuisine table and other modifications to work with muliple cuisines.
Adds a separate table similar to the category table, but for cuisines.
This is a first pass at adding all the functionality.

Adds cuisine values when new recipe is added or modified.  Searched for
category and added a cuisine function where appropriate.

No ability to delete cuisines yet.

Added a way to move cuisines from recipe table to cuisine table. This is
achieved using the migrate_cuisines_to_cuisine_table() method from the
RecData class in src/gourmand/backends.db.py.

It is accomplished without introducing duplicate cuisines and only
bothers with recipes which are not deleted.  Also, once the cuisine
information has been transferred over it sets it to NULL in the
recipe table.

Added conditional so automatic migration does not occur during testing

Added delete_cuisine() method to db.py to delete a single cuisine

Added tests for cuisine table and allow deletion of cuisine through modify_rec