Modified views associated with the change in cuisine information.
The previous commit allows for multiple cuisines to be associated
with a recipe by saving that information in a separate table. This
commit updates the UI files to add a cuisine tab on the recipe
editor notebook.  This separate tab allows one to add and delete
multiple cuisines.

Additionally, the taxonomy of cuisines has been simplified.  Instead
of Asian/Thai and Asian/Indian (and similar categorization) the names
have been changed to Thai and Indian.