Allow timer highlighting to work with time spans (#233)
Closes #231.

Used the time string information that was already parsed when
making the link in the first place, rather than passing the
time string to a different regex parsing function.  For example,
the actual text in a recipe reads '35 to 40 minutes' and
currently that is what is passed.  The value stored in
markup_dict['35 to 40 minutes'] is '35 minutes'.  Part of the
code changes in this commit is passing '35 minutes' instead of
the entire phrase '35 to 40 minutes'.

Moved the time_matcher regex setup to convert.py and imported that
into timeScanner.py vs multiple regexes.

Removed timestring_to_seconds_old function since it is never used
based on a text search of the whole project.