Fix XML around instructions and notes (modifications). (#223)
Closes #77.

Gourmand double escapes xml special characters
&, < and > when exporting.  This causes an ampersand (&) in the text
to become &amp;amp;.  A similar issue happens with > and <.  To combat
this an extra call to xml.sax.saxutils.unescape() was used in the
exporter class private method _write_text_.

The opposite issue is present when importing.  The special characters
needed to be escaped.  To achieve this the importer.Importer.__init__()
method is given a default do_markup value of True.  This seems to cause
the special symbols to be escaped properly when importing and saving.

These escaping issues seemed to only be present in the instructions and
modification sections of the recipe, which indicates the solution could
be found by changing code in those code paths.

The previous commit worked to allow <, > and & to be imported and
exported since they commonly interfere with XML parsing.  Unfortunately,
it was found the fix for that broke saved formatting information
which is communicated through <b>, <u> and <i> tags, and possibly
others.

The errors with handling formatting tags and special characters at the
same time stem from improper escaping during both the export and import
process.  To fix with with exporting a single extra escape line was added
to the exporter class.  This extra escape is only applied to the text
portion of the XML document and not the tags themselves and it has the
effect of preserving the relative escape level (for lack of a better term).

During the import process both an unescape and an escape portion of the
code applied to 'instructions' and 'modifications' had to be removed.
While it might seem like these cancel out, they were applied in such a
way to abolish the relative escape level, and removing these again helps
preserve the relative escape level during the import process.

See the thread in issue #77 for more context.

An extra file was added at tests/recipe_files/special_chars.grmt.
This contains the gourmetDoc xml file for a simple recipe with only
a title and instructions with complicated characters and format.

For some reason re.match did not work, probably due to the special
characters or newlines.  For this reason I added a couple functions
modelled off of the other ones, to import the file and make sure
the string strictly matches what is expected.